Move ros::names implementation to own file
